+++ packages/editor/src/store/test/selectors.js
@@ -89,7 +89,6 @@ const {
 	didPostSaveRequestFail,
 	getSuggestedPostFormat,
 	getEditedPostContent,
-	getNotices,
 	getReusableBlock,
 	isSavingReusableBlock,
 	isFetchingReusableBlock,
@@ -111,6 +110,7 @@ const {
 	INSERTER_UTILITY_HIGH,
 	INSERTER_UTILITY_MEDIUM,
 	INSERTER_UTILITY_LOW,
+	isPostSavingLocked,
 } = selectors;
 
 describe( 'selectors', () => {
@@ -888,6 +888,28 @@ describe( 'selectors', () => {
 			};
 
 			expect( isEditedPostPublishable( state ) ).toBe( true );
+		} );
+	} );
+
+	describe( 'isPostSavingLocked', () => {
+		it( 'should return true if the post has postSavingLocks', () => {
+			const state = {
+				postSavingLock: [ { 1: true } ],
+				currentPost: {},
+				saving: {},
+			};
+
+			expect( isPostSavingLocked( state ) ).toBe( true );
+		} );
+
+		it( 'should return false if the post has no postSavingLocks', () => {
+			const state = {
+				postSavingLock: [],
+				currentPost: {},
+				saving: {},
+			};
+
+			expect( isPostSavingLocked( state ) ).toBe( false );
 		} );
 	} );
 
@@ -3255,19 +3277,6 @@ describe( 'selectors', () => {
 			const content = getEditedPostContent( state );
 
 			expect( content ).toBe( '<!-- wp:default {\"modified\":true} /-->' );
-		} );
-	} );
-
-	describe( 'getNotices', () => {
-		it( 'should return the notices array', () => {
-			const state = {
-				notices: [
-					{ id: 'b', content: 'Post saved' },
-					{ id: 'a', content: 'Error saving' },
-				],
-			};
-
-			expect( getNotices( state ) ).toEqual( state.notices );
 		} );
 	} );