+++ packages/blocks/CHANGELOG.md
@@ -1,4 +1,10 @@
-## 4.1.0 (Unreleased)
+## 5.0.0 (Unreleased)
+
+### Breaking Changes
+
+- Attribute type coercion has been removed. Omit the source to preserve type via serialized comment demarcation.
+- `setUnknownTypeHandlerName` has been removed. Please use `setFreeformContentHandlerName` and `setUnregisteredTypeHandlerName` instead.
+- `getUnknownTypeHandlerName` has been removed. Please use `getFreeformContentHandlerName` and `getUnregisteredTypeHandlerName` instead.
 
 ### New Feature